- FC and FlowControl (UFC/MFC) don't need to handle a CONFIG(frag_size) from below, as it only makes sense (at least currently) to have FRAG2 sit above a flow control protocol
- FRAG/FRAG2 don't need to handle their own CONFIG event
JIRA: https://issues.jboss.org/browse/JGRP-1712

+++ src/org/jgroups/protocols/RATE_LIMITER.java
@@ -141,7 +141,7 @@ public class RATE_LIMITER extends Protocol {
             Map<String,Object> map=(Map<String, Object>)evt.getArg();
             Integer tmp=map != null? (Integer)map.get("frag_size") : null;
             if(tmp != null)
-                frag_size=tmp.intValue();
+                frag_size=tmp;
             if(frag_size > 0) {
                 if(max_bytes % frag_size != 0) {
                     if(log.isWarnEnabled())
